Refactor: repackage and rename to AbstractEntityDetailsDialog
