+++ OpERP/src/main/java/devopsdistilled/operp/client/abstracts/AbstractEntityDetailsDialog.java
@@ -1,15 +1,15 @@
-package devopsdistilled.operp.client.items.views;
+package devopsdistilled.operp.client.abstracts;
 
 import javax.swing.JComponent;
 import javax.swing.JDialog;
 
 import devopsdistilled.operp.server.data.entity.Entiti;
 
-public abstract class EntityDetailsDialog<E extends Entiti> {
+public abstract class AbstractEntityDetailsDialog<E extends Entiti> {
 
 	protected final JDialog dialog;
 
-	public EntityDetailsDialog() {
+	public AbstractEntityDetailsDialog() {
 		dialog = new JDialog();
 		dialog.setDefaultCloseOperation(JDialog.DISPOSE_ON_CLOSE);
 		dialog.setTitle("Details");
